# openapi.yml
openapi: 3.0.1
info:
  title: Nunito Backend API
  version: "1.0.0"
servers:
- url: http://localhost:8080
  description: Servidor local
...
paths:
  /api/courses:
    get:
      tags:
      - courses
      summary: Listar cursos del profesor
      security:
      - bearerAuth: []
      parameters:
      - name: teacherId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Lista de cursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      tags:
      - courses
      summary: Crear un nuevo curso
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Curso creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /api/courses/{courseId}:
    get:
      tags:
      - courses
      summary: Obtener detalle de un curso
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Detalle del curso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      tags:
      - courses
      summary: Actualizar un curso
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Curso actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      tags:
      - courses
      summary: Eliminar un curso
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Curso eliminado
  /api/courses/{courseId}/test-suites:
    get:
      tags:
      - test-suites
      summary: Listar conjuntos de preguntas de un curso
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Lista de conjuntos de preguntas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestSuite"
  /api/test-suites:
    post:
      tags:
      - test-suites
      summary: Crear un nuevo conjunto de preguntas
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestSuiteRequest"
      responses:
        "201":
          description: Conjunto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSuite"
  /api/test-suites/{testSuiteId}:
    get:
      tags:
      - test-suites
      summary: Obtener detalle de un conjunto de preguntas
      security:
      - bearerAuth: []
      parameters:
      - name: testSuiteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Detalle del conjunto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSuite"
    patch:
      tags:
      - test-suites
      summary: Actualizar un conjunto de preguntas
      security:
      - bearerAuth: []
      parameters:
      - name: testSuiteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestSuiteRequest"
      responses:
        "200":
          description: Conjunto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSuite"
    delete:
      tags:
      - test-suites
      summary: Eliminar un conjunto de preguntas
      security:
      - bearerAuth: []
      parameters:
      - name: testSuiteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Conjunto eliminado
  /api/test-suites/{testSuiteId}/questions:
    get:
      tags:
      - questions
      summary: Listar preguntas de un conjunto
      security:
      - bearerAuth: []
      parameters:
      - name: testSuiteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: gameType
        in: query
        required: false
        schema:
          $ref: "#/components/schemas/GameId"
      responses:
        "200":
          description: Lista de preguntas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      tags:
      - questions
      summary: Crear una nueva pregunta en un conjunto
      security:
      - bearerAuth: []
      parameters:
      - name: testSuiteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Pregunta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /api/questions/{questionId}:
    get:
      tags:
      - questions
      summary: Obtener detalle de una pregunta
      security:
      - bearerAuth: []
      parameters:
      - name: questionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Detalle de la pregunta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      tags:
      - questions
      summary: Actualizar una pregunta
      security:
      - bearerAuth: []
      parameters:
      - name: questionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "200":
          description: Pregunta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      tags:
      - questions
      summary: Eliminar una pregunta
      security:
      - bearerAuth: []
      parameters:
      - name: questionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Pregunta eliminada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GameId:
      type: string
      enum:
      - image-word
      - syllable-count
      - rhyme-identification
      - audio-recognition
...
    Course:
      type: object
      required:
      - id
      - name
      - teacherId
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        teacherId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCourseRequest:
      type: object
      required:
      - name
      - teacherId
      properties:
        name:
          type: string
        description:
          type: string
        teacherId:
          type: string
          format: uuid
    UpdateCourseRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    TestSuite:
      type: object
      required:
      - id
      - name
      - courseId
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        courseId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTestSuiteRequest:
      type: object
      required:
      - name
      - courseId
      properties:
        name:
          type: string
        description:
          type: string
        courseId:
          type: string
          format: uuid
    UpdateTestSuiteRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    Question:
      type: object
      required:
      - id
      - text
      - type
      - testSuiteId
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        type:
          $ref: "#/components/schemas/GameId"
        options:
          type: object
          description: "JSON flexible con las opciones de la pregunta"
        correctAnswer:
          type: string
        testSuiteId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    CreateQuestionRequest:
      type: object
      required:
      - text
      - type
      - testSuiteId
      properties:
        text:
          type: string
        type:
          $ref: "#/components/schemas/GameId"
        options:
          type: object
        correctAnswer:
          type: string
        testSuiteId:
          type: string
          format: uuid
...